+++ Planetario/Planetario/Models/PreguntasFrecuentes.cs
@@ -5,7 +5,16 @@ using System.Web;
 
 namespace Planetario.Models
 {
-    public class PreguntasFrecuentes
+    public class PreguntasFrecuentesModel
     {
+        public int idPregunta { get; set; }
+
+        public string categoriaPregunta { get; set; }
+
+        public string topicoPregunta { get; set; }
+
+        public string pregunta { get; set; }
+
+        public string respuesta { get; set; }
     }
 }
