+++ Planetario/Planetario/Views/Shared/_Layout.cshtml
@@ -24,7 +24,7 @@
                     <li>@Html.ActionLink("Inicio", "Index", "Home")</li>
                     <li>@Html.ActionLink("Acerca de", "About", "Home")</li>
                     <li>@Html.ActionLink("Contacto", "Contact", "Home")</li>
-                    <li>@Html.ActionLink("Preguntas Frecuentes", "preguntasFrecuentes", "PreguntasFrecuentes")</li>
+                    <li>@Html.ActionLink("Preguntas Frecuentes", "PreguntasFrecuentes", "PreguntasFrecuentes")</li>
                 </ul>
             </div>
         </div>
