+++ Planetario/Planetario/Views/Web.config
@@ -40,4 +40,7 @@
       </assemblies>
     </compilation>
   </system.web>
+	<connectionStrings>
+		<add name="PlanetarioConnection" connectionString="Data Source=172.16.202.45;Initial Catalog=BD_Grupo03;Persist Security Info=True;User ID=BD_Grupo03;Password=Grupo03DB" />
+	</connectionStrings>
 </configuration>
